feat: seperate candlestick data from marketContext
